renamed todoStore to taskStore
