+++ src/store/store.ts
@@ -1,5 +1,5 @@
 import { authState, login, logout, register } from "./authStore";
-import { taskState, fetchTasks, addTask, updateTask, deleteTask } from "./todoStore";
+import { taskStore, fetchTasks, addTask, updateTask, deleteTask } from "./taskStore";
 
 export const store = {
     auth: {
@@ -8,8 +8,8 @@ export const store = {
         logout,
         register,
     },
-    todo: {
-        state: taskState,
+    tasks: {
+        state: taskStore,
         fetchTasks,
         addTask,
         updateTask,
